first manager method tested
