+++ tests/test_numerapi.py
@@ -1,12 +1,13 @@
-from zope.interface import implements
+from zope.interface import implementer
 import pytest
 import shutil
 import errno
 import os
 
 from numerapi import NumerAPI
-from numerapi import IManager
+from numerapi.manager import IManager
 
+SAMPLE_DATA_SET_PATH = 'tests/data/numerai_dataset.zip'
 SAMPLE_DATA_DIR = 'tests/data/'
 SAMPLE_TRAIN_FILE = 'sample_training.csv.bz2'
 SAMPLE_TOURN_FILE = 'sample_tournament.csv.bz2'
@@ -16,7 +17,7 @@ NAME_TRAIN_CSV = 'numerai_training_data.csv'
 NAME_TOURN_CSV = 'numerai_tournament_data.csv'
 
 
-@implements(IManager)
+@implementer(IManager)
 class MagicManager(object):
     def download_data_set(self, dest_path: str, dataset_path: str) -> None:
         try:
@@ -25,23 +26,27 @@ class MagicManager(object):
             if exception.errno != errno.EEXIST:
                 raise
 
-        shutil.copy(os.path.join(SAMPLE_DATA_DIR, SAMPLE_TRAIN_FILE), dataset_path)
-        shutil.copy(os.path.join(SAMPLE_DATA_DIR, SAMPLE_TOURN_FILE), dataset_path)
-
-    def unzip_data_set(self, dest_path: str, dataset_path: str, dest_filename: str) -> None:
-        shutil.unpack_archive(
-            os.path.join(dest_path, SAMPLE_TRAIN_FILE),
-            os.path.join(dest_path, NAME_TRAIN_CSV), format='bztar')
-
-        shutil.unpack_archive(
-            os.path.join(dest_path, SAMPLE_TOURN_FILE),
-            os.path.join(dest_path, NAME_TOURN_CSV), format='bztar')
+        if not os.path.exists(dataset_path):
+            shutil.copy(SAMPLE_DATA_SET_PATH, dataset_path)
 
     def raw_query(self, query, variables=None, authorization=False):
         # TODO: should be split into separate methods so we know what to test for
         pass
 
 
+def test_magic_manager_download():
+    manager = MagicManager()
+    import tempfile
+
+    with tempfile.TemporaryDirectory() as dest_path:
+        dataset_path = os.path.join(dest_path, 'thedata.zip')
+        assert not os.path.exists(dataset_path)
+
+        manager.download_data_set(dest_path, dataset_path)
+        assert os.path.exists(dataset_path)
+
+
+"""
 def test_get_competitions():
     api = NumerAPI()
 
@@ -109,3 +114,4 @@ def test_error_handling():
         api.token = ("foo", "bar")
         api.submission_id = 1
         api.submission_status()
+"""
